+++ morpheus/changelog.md
@@ -1,3 +1,4 @@
 # 0.0.1 to 0.0.2
 - interpreter can parse input Objects and Arrays (all input values)
 - fix: argument supports nonNull and List Types in introspection
+- add class Interpreter and instances for: ByteString , Text , Lazy ByteString, Lazy Text